Setup the initial file
This file will create a venv and let the user to select the location too. After that it will install the req txt and all the libs and module that are required to run this engine file.

i already mentioned this file into req txt to call this as the initial file to start the process.